+++ src/test/kotlin/com/flab/inqueue/security/hmacsinature/HmacSignatureSecurityTest.kt
@@ -1,14 +1,12 @@
 package com.flab.inqueue.security.hmacsinature
 
 import com.flab.inqueue.AcceptanceTest
-import com.flab.inqueue.domain.dto.AuthRequest
 import com.flab.inqueue.domain.member.entity.Member
 import com.flab.inqueue.domain.member.entity.MemberKey
 import com.flab.inqueue.domain.member.repository.MemberRepository
 import com.flab.inqueue.domain.member.utils.memberkeygenrator.MemberKeyGenerator
 import com.flab.inqueue.security.common.Role
 import com.flab.inqueue.security.hmacsinature.utils.EncryptionUtil
-import com.github.dockerjava.zerodep.shaded.org.apache.commons.codec.binary.Base64
 import jakarta.transaction.Transactional
 import org.hamcrest.Matchers
 import org.junit.jupiter.api.BeforeEach
@@ -19,9 +17,6 @@ import org.springframework.boot.test.web.server.LocalServerPort
 import org.springframework.http.HttpHeaders
 import org.springframework.http.HttpStatus
 import org.springframework.http.MediaType
-import javax.crypto.Mac
-import javax.crypto.spec.SecretKeySpec
-
 
 class HmacSignatureSecurityTest : AcceptanceTest() {
 
@@ -78,7 +73,7 @@ class HmacSignatureSecurityTest : AcceptanceTest() {
         given.log().all()
             .header(
                 HttpHeaders.AUTHORIZATION,
-                createAuthorization(
+                createHmacAuthorizationHeader(
                     notEncryptedUserMemberKey.clientId,
                     createHmacSignature(hmacSignaturePayloadWithUser, notEncryptedUserMemberKey.clientSecret)
                 )
@@ -109,7 +104,7 @@ class HmacSignatureSecurityTest : AcceptanceTest() {
         given.log().all()
             .header(
                 HttpHeaders.AUTHORIZATION,
-                createAuthorization(
+                createHmacAuthorizationHeader(
                     notEncryptedUserMemberKey.clientId,
                     createHmacSignature(hmacSignaturePayloadWithUser, anotherMemberKey.clientSecret)
                 )
@@ -129,21 +124,16 @@ class HmacSignatureSecurityTest : AcceptanceTest() {
     @Test
     @DisplayName("clientId를 찾을 수 없는 경우, Hmac Authentication 실패")
     fun hmac_authentication_fail2() {
-        val eventId = "estEventId"
-        val userId = "testUserId"
-
         val anotherMemberKey = memberKeyGenerator.generate()
-        val authRequest = AuthRequest(eventId, userId)
 
         given.log().all()
             .header(
                 HttpHeaders.AUTHORIZATION,
-                createAuthorization(
+                createHmacAuthorizationHeader(
                     anotherMemberKey.clientId,
                     createHmacSignature(hmacSignaturePayloadWithUser, anotherMemberKey.clientSecret)
                 )
             )
-            .body(authRequest)
             .contentType(MediaType.APPLICATION_JSON_VALUE).
         `when`()
             .post(HMAC_SECURITY_TEST_URI).
@@ -159,13 +149,7 @@ class HmacSignatureSecurityTest : AcceptanceTest() {
     @Test
     @DisplayName("AUTHORIZATION 헤더가 없는 경우, Hmac Authentication 실패")
     fun hmac_authentication_fail3() {
-        val eventId = "estEventId"
-        val userId = "testUserId"
-
-        val authRequest = AuthRequest(eventId, userId)
-
         given.given().log().all()
-            .body(authRequest)
             .contentType(MediaType.APPLICATION_JSON_VALUE).
         `when`()
             .post(HMAC_SECURITY_TEST_URI).
@@ -184,7 +168,7 @@ class HmacSignatureSecurityTest : AcceptanceTest() {
         given.log().all()
             .header(
                 HttpHeaders.AUTHORIZATION,
-                createAuthorization(
+                createHmacAuthorizationHeader(
                     notEncryptedAdminMemberKey.clientId,
                     createHmacSignature(hmacSignaturePayloadWithAdmin, notEncryptedAdminMemberKey.clientSecret)
                 )
@@ -213,7 +197,7 @@ class HmacSignatureSecurityTest : AcceptanceTest() {
         given.log().all()
             .header(
                 HttpHeaders.AUTHORIZATION,
-                createAuthorization(
+                createHmacAuthorizationHeader(
                     notEncryptedUserMemberKey.clientId,
                     createHmacSignature(hmacSignaturePayloadWithAdmin, notEncryptedUserMemberKey.clientSecret)
                 )
@@ -228,16 +212,5 @@ class HmacSignatureSecurityTest : AcceptanceTest() {
             .body("timestamp", Matchers.notNullValue())
             .body("status", Matchers.notNullValue())
             .body("path", Matchers.notNullValue())
-    }
-
-    private fun createHmacSignature(payLoad: String, clientSecret: String): String {
-        val sha256HMAC = Mac.getInstance("HmacSHA256")
-        val secretKey = SecretKeySpec(clientSecret.toByteArray(), "HmacSHA256")
-        sha256HMAC.init(secretKey)
-        return Base64.encodeBase64String(sha256HMAC.doFinal(payLoad.toByteArray()))
-    }
-
-    private fun createAuthorization(clientId: String, hmacSignature: String): String {
-        return "$clientId:$hmacSignature"
     }
 }
